WebsiteProject
This website was created as a personal project to utilize the knowledge of programming languages and not limited to such as HTML, CSS, JavaScript and Microsoft SQL Database. The website's objective is to introduce the importance of self-awarenss, self-help and healthy lifestyle and allow the registered user to select the package to further expand the knowledge and get it. The website wasnt extended to e-commerc.